fix: add missing tooltips for resource/array props
These tooltips are super useful! This just adds two that are missing.

I've been consistently clicking the 'reference' icon instead of the
'resource' one, but that's me being dumb.

Also, the 'clear' icon irks me - is it supposed to be a broom? I was
bothered until I realized it's also used in Godot's output UI, so I
suppose it's fine - but I think the trashcan (used in a few places)
might read a bit better for deleting things.